backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "sqlmodel>=0.0.31",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.10.1",
    "passlib>=1.7.4",
    "python-jose[bcrypt]>=3.5.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "bcrypt>=5.0.0",
]
